fix waste reminder complexity gate
